chore: more cleanup
