+++ rust/Cargo.toml
@@ -115,6 +115,5 @@ pretty_assertions = "1"
 tempdir = "0"
 tempfile = "3"
 maplit = { version = "1" }
-anyhow = "1"
 rand = "0.8"
 dotenv = "*"
