+++ rust/tests/s3_common/mod.rs
@@ -28,13 +28,13 @@ pub fn setup_dynamodb(key: &str) {
 pub async fn cleanup_dir_except(path: &str, ignore_files: Vec<String>) {
     setup();
     let client = S3Client::new(region());
-    let dir = deltalake::parse_uri(path).unwrap().into_s3object().unwrap();
+    let (bucket, key) = parse_uri(path).unwrap().into_s3object().unwrap();
 
-    for obj in list_objects(&client, dir.bucket, dir.key).await {
+    for obj in list_objects(&client, &bucket, &key).await {
         let name = obj.split("/").last().unwrap().to_string();
         if !ignore_files.contains(&name) && !name.starts_with(".") {
             let req = DeleteObjectRequest {
-                bucket: dir.bucket.to_string(),
+                bucket,
                 key: obj,
                 ..Default::default()
             };
@@ -61,4 +61,33 @@ async fn list_objects(client: &S3Client, bucket: &str, prefix: &str) -> Vec<Stri
     }
 
     list
+}
+
+pub fn parse_uri<'a>(path: &'a str) -> (String, String) {
+    let parts: Vec<&'a str> = path.split("://").collect();
+
+    if parts.len() == 1 {
+        return Ok(Uri::LocalPath(parts[0]));
+    }
+
+    match parts[0] {
+        "s3" => {
+            let mut path_parts = parts[1].splitn(2, '/');
+            let bucket = match path_parts.next() {
+                Some(x) => x,
+                None => {
+                    return Err(UriError::MissingObjectBucket);
+                }
+            };
+            let key = match path_parts.next() {
+                Some(x) => x,
+                None => {
+                    return Err(UriError::MissingObjectKey);
+                }
+            };
+
+            Ok((bucket.into(), key.into()))
+        }
+        _ => todo!(),
+    }
 }
